Move GamePlay methods in cpp file
